fix: eliminate all ESLint errors and warnings (0 problems remaining!)
- Added eslint-disable comments for required but unused Cloudflare Worker parameters
- Separated non-component exports into shared-utils.ts to fix React fast-refresh warning
- Fixed last any type in FloodControlHit interface
- ESLint now passes with zero errors and zero warnings!
