+++ lint.txt
@@ -1,13 +1,6 @@
 
-/Users/arke/Projects/bettergov/functions/api/forex.ts
-   47:31  error  '_env' is defined but never used  @typescript-eslint/no-unused-vars
-  246:3   error  '_ctx' is defined but never used  @typescript-eslint/no-unused-vars
+/Users/arke/Projects/bettergov/src/pages/flood-control-projects/shared-utils.ts
+  40:37  error  Unexpected any. Specify a different type  @typescript-eslint/no-explicit-any
 
-/Users/arke/Projects/bettergov/functions/api/weather.ts
-  316:3  error  '_ctx' is defined but never used  @typescript-eslint/no-unused-vars
-
-/Users/arke/Projects/bettergov/src/pages/flood-control-projects/shared-components.tsx
-  263:14  warning  Fast refresh only works when a file only exports components. Use a new file to share constants or functions between components  react-refresh/only-export-components
-
-✖ 4 problems (3 errors, 1 warning)
+✖ 1 problem (1 error, 0 warnings)
