+++ src/pages/flood-control-projects/shared-components.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 import { ChevronLeft } from 'lucide-react';
 import { ScrollArea } from '../../components/ui/ScrollArea';
+import type { FilterState, FloodControlProject } from './shared-utils';
 
 // Define types
 export interface DataItem {
@@ -8,50 +9,9 @@ export interface DataItem {
   count: number;
 }
 
-export type FloodControlHit = {
-  GlobalID?: string;
-  objectID?: string;
-  ProjectDescription?: string;
-  InfraYear?: string;
-  Region?: string;
-  Province?: string;
-  Municipality?: string;
-  TypeofWork?: string;
-  Contractor?: string;
-  ContractCost?: string;
-  DistrictEngineeringOffice?: string;
-  LegislativeDistrict?: string;
-};
-
-export type FloodControlProject = {
-  ProjectDescription?: string;
-  Municipality?: string;
-  Region?: string;
-  Province?: string;
-  ContractID?: string;
-  ProjectID?: string;
-  ContractCost?: number;
-  TypeofWork?: string;
-  LegislativeDistrict?: string;
-  DistrictEngineeringOffice?: string;
-  InfraYear?: string;
-  Contractor?: string;
-  slug?: string;
-};
-
 export interface HitProps {
   hit: FloodControlProject;
 }
-
-export type FilterState = {
-  InfraYear: string;
-  Region: string;
-  Province: string;
-  TypeofWork: string;
-  DistrictEngineeringOffice: string;
-  LegislativeDistrict: string;
-  [key: string]: string;
-};
 
 // Reusable FilterDropdown component
 export interface FilterDropdownProps {
@@ -257,46 +217,4 @@ export const ResultsStatistics: React.FC<{
       </div>
     </div>
   );
-};
-
-// Utility function to build filter string
-export const buildFilterString = (filters: FilterState): string => {
-  // Start with an empty array - we'll add filters as needed
-  const filterStrings: string[] = [];
-
-  // Always filter by type
-  filterStrings.push('type = "flood_control"');
-
-  if (filters.InfraYear && filters.InfraYear.trim()) {
-    filterStrings.push(`FundingYear = ${filters.InfraYear.trim()}`);
-  }
-
-  if (filters.Region && filters.Region.trim()) {
-    filterStrings.push(`Region = "${filters.Region.trim()}"`);
-  }
-
-  if (filters.Province && filters.Province.trim()) {
-    filterStrings.push(`Province = "${filters.Province.trim()}"`);
-  }
-
-  if (filters.TypeofWork && filters.TypeofWork.trim()) {
-    filterStrings.push(`TypeofWork = "${filters.TypeofWork.trim()}"`);
-  }
-
-  if (
-    filters.DistrictEngineeringOffice &&
-    filters.DistrictEngineeringOffice.trim()
-  ) {
-    filterStrings.push(
-      `DistrictEngineeringOffice = "${filters.DistrictEngineeringOffice.trim()}"`
-    );
-  }
-
-  if (filters.LegislativeDistrict && filters.LegislativeDistrict.trim()) {
-    filterStrings.push(
-      `LegislativeDistrict = "${filters.LegislativeDistrict.trim()}"`
-    );
-  }
-
-  return filterStrings.length > 0 ? filterStrings.join(' AND ') : '';
 };
